requirements.txt
Flask
requests
imbox
python-dotenv
mysql-connector-python
